+++ src/hooks/interact/useListItemMultDelete.ts
@@ -1,12 +1,16 @@
 import { useMutation } from '@tanstack/react-query'
 import InteractAPIClient from '@/services/api-interact'
-import ListItemTuples from '@/entities/ListItemTuples'
+import CollectionItemTuple from '@/entities/CollectionItemTuple'
+import { useAuth } from '@/AuthContext'
 
 const apiClient = new InteractAPIClient('collections')
 
 const useListItemMultDelete = () => {
+  const { user } = useAuth()
+
   return useMutation({
-    mutationFn: async (listItemTuples: ListItemTuples[]) => {
+    mutationFn: async (listItemTuples: CollectionItemTuple[]) => {
+      if (!user) return
       const responses = await Promise.allSettled(
         listItemTuples.map(({ listId, listItemId }) =>
           apiClient.deleteListItem(
